Adding some basic types: U32, S32, etc.. The number in the type name indicates the type's size in bits. The letter stands for unsigned, signed, or floating-point.
I've included some template magic to issue a compiler error if any of these types are not the size they say they are.

I've been using these types daily for many years, and I find them really nice to work with, easy to type, easy to read, and generally unambiguous.

git-svn-id: http://oolongengine.googlecode.com/svn/trunk@73 7fa7efda-f44a-0410-a86a-c1fddb4bcab7