Adding a psuedorandom number generator wrapper abstract base class.
It's useful in that you can conveniently have multiple random number generator objects, each with it's own state, and each producing the same stream of pseudorandom numbers.

A default random number generator based on rand() is provided. The provided generator is not very good, but dropping a better RNG in is a fairly trivial task.

git-svn-id: http://oolongengine.googlecode.com/svn/trunk@74 7fa7efda-f44a-0410-a86a-c1fddb4bcab7