Okay, so the Skeleton2 project now compiles and runs on the iPhone device 3.1.
A few words about this (extraordinarily) fun process:

It seems that the memory manager complains because of the way that resources are being created and deleted.  When opening shader files and running them through their own separate files, this results in a few calls to fopen and fclose.  On the call to fopen, nothing is malloc'd.  On the call to fclose, the memory manager reports a call to delete, which (naturally) is deleting something that was never created.  Seeing as how fopen and fclose are functions that we do not touch, I find no fault with the memory manager or the sample code.

This is by no means the definitive answer, simply what I have discovered and noticed, and if somebody else has a better idea of what is going on or a better fix, let's go with that? =)

In order to make this run, I have edited the project.pbxproj to add the PreProcessor Macro _NOMEMORYMANAGER, which disables the memory manager (for the time being).  The program will not compile for iPhone OS 3.0 (It never did), and it is now running on iPhone 3.1 Beta 1.  Supposedly beta 2 is out, this might fix things?

In addition, there is a constant error that makes itself known in the debug console, that it can't run a certain load instruction?  A little bit of googling make it seem like this is a known issue with 3.1, so hopefully it will be resolved soon.

Alan Green

git-svn-id: http://oolongengine.googlecode.com/svn/trunk@144 7fa7efda-f44a-0410-a86a-c1fddb4bcab7